update monitor: refactoring, fix page query param delay problem
